Removes components
