+++ app/client/src/ee/components/editorComponents/GPT/index.tsx
@@ -1,0 +1,1 @@
+export * from "ce/components/editorComponents/GPT";
